+++ server/src/controllers/group.controller.ts
@@ -9,28 +9,20 @@ import {
 } from '../loaders/orm.loader';
 import { ChattingChannel, MeetingChannel, Workgroup } from '../db/entities';
 import { ChannelType } from '../types/ChannelType';
-import { Repository } from 'typeorm';
+import { GROUP_MSG } from '../messages';
 
 const nullCheck = (data) => data !== undefined && data !== null && data !== '';
 const encodeBase64 = (str: string): string => Buffer.from(str, 'binary').toString('base64');
 
-const MSG = {
-  needGroupName: '그룹명이 누락되었습니다.',
-  userNotFound: '존재하지 않는 회원입니다.',
-  groupNotFound: '존재하지 않는 그룹입니다.',
-  wrongChannelType: '올바르지 않은 채널 타입입니다.',
-  alreadyJoined: '이미 그룹에 가입된 사용자입니다.',
-  channelNameEmpty: '채널 이름을 입력해주세요',
-};
 const DEFAULT_CHANNEL_NAME = 'general';
 
 const createGroup = async (req: Request, res: Response, next: NextFunction) => {
   const { groupName, groupThumbnail } = req.body;
   try {
-    if (!nullCheck(groupName)) return res.status(400).send(MSG.needGroupName);
+    if (!nullCheck(groupName)) return res.status(400).send(GROUP_MSG.needGroupName);
     const leaderID = req.session.userID;
     const leader = await userRepository.findOne({ where: { id: leaderID } });
-    if (!leader) return res.status(400).send(MSG.userNotFound);
+    if (!leader) return res.status(400).send(GROUP_MSG.userNotFound);
 
     const newGroup = new Workgroup();
     newGroup.name = groupName;
@@ -77,7 +69,7 @@ const getGroupMembers = async (req: Request, res: Response, next: NextFunction) 
   const { id } = req.params;
   try {
     const group = await groupRepository.findOne({ where: { id: id } });
-    if (!group) return res.status(400).send(MSG.groupNotFound);
+    if (!group) return res.status(400).send(GROUP_MSG.groupNotFound);
 
     const members = await groupMemberRepository.findUsersByGroupID(group.id);
 
@@ -92,11 +84,12 @@ const createChannel = async (req: Request, res: Response, next: NextFunction) =>
   const { channelType, channelName } = req.body;
 
   try {
-    if (!channelName || !channelName.trim()) return res.status(400).send(MSG.channelNameEmpty);
+    if (!channelName || !channelName.trim())
+      return res.status(400).send(GROUP_MSG.channelNameEmpty);
     const group = await groupRepository.findOne({ where: { id: id } });
-    if (!group) return res.status(400).send(MSG.groupNotFound);
+    if (!group) return res.status(400).send(GROUP_MSG.groupNotFound);
     if (!Object.values(ChannelType).includes(channelType))
-      return res.status(400).send(MSG.wrongChannelType);
+      return res.status(400).send(GROUP_MSG.wrongChannelType);
 
     const newChannel =
       channelType === ChannelType.chatting ? new ChattingChannel() : new MeetingChannel();
@@ -123,7 +116,7 @@ const joinGroup = async (req: Request, res: Response, next: NextFunction) => {
       where: { code: groupCode },
       relations: ['meetingChannels', 'chattingChannels'],
     });
-    if (!group) return res.status(400).send(MSG.groupNotFound);
+    if (!group) return res.status(400).send(GROUP_MSG.groupNotFound);
 
     const relation = await groupMemberRepository
       .createQueryBuilder('group_member')
@@ -133,7 +126,7 @@ const joinGroup = async (req: Request, res: Response, next: NextFunction) => {
       })
       .getOne();
 
-    if (relation) return res.status(400).send(MSG.alreadyJoined);
+    if (relation) return res.status(400).send(GROUP_MSG.alreadyJoined);
 
     const now = new Date();
     await groupMemberRepository.insert({ group: group, user: user, lastAccessTime: now });
@@ -171,7 +164,7 @@ const deleteChannel = async (req: Request, res: Response, next: NextFunction) =>
     if (!group) return res.status(400).send('존재하지 않는 그룹 아이디입니다.');
     if (group.leader.id !== userID) return res.status(400).send('채널 삭제 권한이 없습니다.');
     if (!['meeting', 'chatting'].includes(channelType))
-      return res.status(400).send(MSG.wrongChannelType);
+      return res.status(400).send(GROUP_MSG.wrongChannelType);
 
     if (channelType === 'meeting') {
       const channel = await meetingChannelRepository.findOne({
