+++ server/src/messages/channel.mesage.ts
@@ -1,0 +1,6 @@
+export const createChatMSG = {
+  userNotFound: '존재하지 않는 사용자 입니다.',
+  chatNotFound: '존재하지 않는 텍스트 입니다.',
+  emptyChat: '채팅을 입력해 주세요.',
+  success: '스레드 전송 성공!',
+};
