+++ server/src/messages/chat.message.ts
@@ -1,0 +1,6 @@
+export const handleReactionMSG = {
+  userNotFound: '존재하지 않는 사용자 입니다.',
+  chatNotFound: '존재하지 않는 텍스트 입니다.',
+  deleteReactionSuccess: '좋아요 지우기 성공!',
+  addReactionSuccess: '좋아요 만들기 성공!',
+};
