+++ server/src/messages/group.message.ts
@@ -1,0 +1,8 @@
+export const GROUP_MSG = {
+  needGroupName: '그룹명이 누락되었습니다.',
+  userNotFound: '존재하지 않는 회원입니다.',
+  groupNotFound: '존재하지 않는 그룹입니다.',
+  wrongChannelType: '올바르지 않은 채널 타입입니다.',
+  alreadyJoined: '이미 그룹에 가입된 사용자입니다.',
+  channelNameEmpty: '채널 이름을 입력해주세요',
+};
