+++ server/src/messages/index.ts
@@ -1,1 +1,4 @@
 export * from './user.message';
+export * from './chat.message';
+export * from './channel.mesage';
+export * from './group.message';
